requirements.txt
gitpython>=3.1.0
pytz>=2021.1
click>=8.0.0
colorama>=0.4.4
